Deliverable 3 Dev to Main (#52)
* HMTY-33 add MB scraper

* HMTY-33 refactor and merge

* HTMY-33 add ON scraper

* HTMY-33 small change to ON scraper for debugging

* HTMY-33 updated ON scraper to search with CPSO number

* HTMY-33 fix ON scraper to actually run while using license number

* SK scraper

* [HMTY-33] SK scraper added

* Added NB scraper

* [HMTY-11] Merge Client and Build-Deploy Action into Prod (#6) (#27)

* create react app (#3)

* add client dockerfile and docker-compose (#4)

* HMTY-11 Configure Azure Deployment for Client Application (#5)

* Create an auto-deploy file

* HMTY-11 Update workflow to trigger on main

---------


* Update backend/functional/verification/scrapers/scraperON.js


* Update backend/functional/verification/scrapers/scraperON.js


* Update backend/functional/verification/scrapers/scraperON.js


* [HMTY-36] Database Util Refactor and Scaffold Admin Service (#29)

* HMTY-36 Database Util Refactor and Scaffold Admin Service

* HMTY-36 Update Init Comment

* HMTY-33 small fixes

* NL temp commit

* HMTY-33 add verification endpoint and route

* [HMTY-36] Server Admin Get Paginated Prescribers And Patch Prescribers Impl (#31)

* [HMTY-11] Merge Client and Build-Deploy Action into Prod (#6)

* create react app (#3)

* add client dockerfile and docker-compose (#4)

* HMTY-11 Configure Azure Deployment for Client Application (#5)

* Create an auto-deploy file

* HMTY-11 Update workflow to trigger on main

---------


* HMTY-36 Implement Pagination and Get All Prescribers With Search

* HMTY-36 Create Basic Tests

* HMTY-36 Add Patch Prescriber Test

* HMTY-36 Why is Canadian Spelling Like This + Default Search None

* HMTY-36 One More Licence Change

* HMTY-36 Update Licence Number In Prescriber Search Field


* HMTY-36 Why didn't this one get committed

* HMTY-36 Update docs

* HMTY-36 I Hate Canadian English


---------


* HMTY-33 wip

* HMTY-33 gave up trying to mock

* HMTY-33 untested add prescriber stub & update registered

* HMTY-33 forgot await

* HMTY-33 update devcontainer dockerfile to not mess with filesystem

* HMTY-33 simplify run puppeteer different locations

* HMTY-33 fix schema validation

* HMTY-33 tiny refactor

* fix devcontainer

* HMTY-33 only create prescriber stubs for verified

* HMTY-33 ad empty password

* del

* [HMTY-33] Add Quebec (QC) Scraper (#35)

* HMTY-33 add QC scraper

* HTMY 33 fix QC scraper

---------


* HMTY-33-add-PE-scraper (#32)


* removed unused code, made chnages to data entry

* removed an extra line

* Created NL Scraper

* [HMTY-33] Add NS scraper & small refactor to AB, BC scrapers (#22)

* HMTY-33 add NS scraper

* HMTY-33 small refactors to AB & BC scrapers

* HMTY-33 remove typo

* [MISC] Patch Fullstack & deployment (#39)

* complete fullstack docker compose

* update gitignore

* comment out browser spoof

* [HMTY-37] Admin Access/Patch Prescriber Info (#37)

* HMTY-37 Setup Routing for Drawer

* HMTY-37 Working Search and Prescriber Table

* HMTY-37 Add Edit Dialog (non-functional)

* HMTY-37 Refactor Into Component

* HMTY-37 Working Prescriber Patch Dialog

* HMTY-37 Add Refresh Text

* HMTY-37 Add SVG and Page Size Constant

---------


* HMTY-33 don't verify existing prescribers (#43)

* remove extra await


* styling fix

* [HMTY-29] Green Resources map

* added leaflet to package.json

* [HMTY-24] Prescriber-View-all-of-my-logged-prescriptions (#49)

* [HMTY-57] View Prescirber Form Pagination (#33)

* HMTY-57 Commit backend prescriber form pagination

Include class for form, prescriberRoute, test case

* HMTY-57 Update Form.js to Prescription.js

* HMTY-57 Change file and class name

* HMTY-57 Modification based on Kang's comment

* HMTY-58 Add Client Pagination Prescription (#45)

* HMTY-58 Add Client Pagination Prescription

Add routing, api, implementation of paginated prescriber prescription

* HMTY-58 Resolve naming issue

* [HMTY-57] Modification on Paginated Prescriber Prescription Endpoint (#44)

* HMTY-57 Commit backend prescriber form pagination

Include class for form, prescriberRoute, test case

* HMTY-57 Update Form.js to Prescription.js

* HMTY-57 Change file and class name

* HMTY-57 Modification based on Kang's comment

* HMTY-57 Modify Paginated Prescriber Prescription

Modify parameter to enable searching, move fetch function to fetchUtil, modify test regarding changed parameter

* HMTY-57 Add yup object and dbUtils

* HMTY-57 Modify objWithFields() with yup noUnknown

* HMTY-57 Fixed minor issue

* HMTY-24 Fixed minor naming issue

* HMTY-24 Add wrong prescriber test case

* HMTY-24 Add validation on search

* HMTY-24 Remove unused variable

* implemented requested changes

* Add deliverable 3 docs (#51)

---------

+++ backend/tests/utils/dbUtils.js
@@ -1,6 +1,6 @@
 import { MongoClient, Db } from "mongodb";
 import { COLLECTIONS, SERVER } from "../../constants.js"
-import { genericPatient, genericPrescriber, coordinator, assistant } from "./sampleData.js"
+import { genericPatient, genericPrescriber, coordinator, assistant, genericPrescriberPrescription } from "./sampleData.js"
 import bcrypt from "bcryptjs"
 
 const client = new MongoClient(SERVER.MONGO_URL);
@@ -16,12 +16,12 @@ let db;
  * Implicitly creates the necessary collections if they
  * do not already exist.
  */
-export const clearDB = async () => {
-    await db.collection(COLLECTIONS.ADMINS).deleteMany({});
-    await db.collection(COLLECTIONS.PATIENT).deleteMany({});
-    await db.collection(COLLECTIONS.PRESCRIBER).deleteMany({});
-    await db.collection(COLLECTIONS.PATIENT_PRESCRIPTIONS).deleteMany({});
-    await db.collection(COLLECTIONS.PRESCRIBER_PRESCRIPTIONS).deleteMany({});
+export const clearDB = async (clearAdmins = true, clearPrescribers = true, clearPatient = true, clearPatientPrescription = true, clearPrescriberPrescription = true) => {
+    clearAdmins && await db.collection(COLLECTIONS.ADMINS).deleteMany({});
+    clearPrescribers && await db.collection(COLLECTIONS.PRESCRIBER).deleteMany({});
+    clearPatient && await db.collection(COLLECTIONS.PATIENT).deleteMany({});
+    clearPatientPrescription && await db.collection(COLLECTIONS.PATIENT_PRESCRIPTIONS).deleteMany({});
+    clearPrescriberPrescription && await db.collection(COLLECTIONS.PRESCRIBER_PRESCRIPTIONS).deleteMany({});
 }
 
 /**
@@ -30,8 +30,6 @@ export const clearDB = async () => {
 export const connect = async () => {
     try {
         await client.connect();
-        console.log("Connected to MongoDB");
-
         db = client.db(SERVER.DB_NAME);
     } catch (error) {
         console.error("Error connecting to MongoDB:", error);
@@ -82,6 +80,55 @@ export const insertPrescriber = async (modifier = {}) => {
     await db.collection(COLLECTIONS.PRESCRIBER).insertOne(prescriber);
     return prescriber;
 }
+
+/**
+ * Insert numPrescribers prescribers into the db. 
+ * Each prescriber is generated from the generic prescriber format
+ * with incrementing email "prescriber{i}@gmail.com" and 
+ * padded providerCode "ON-JC{i}"
+ * @param {number} numPrescribers 
+ */
+export const insertPrescribers = async (numPrescribers = 20) => {
+    for (let i = 1; i <= numPrescribers; i++) {
+        const modifier = {
+            email: `prescriber${i}@gmail.com`,
+            providerCode: `ON-JC${String(i).padStart(3, '0')}`
+        }
+        await insertPrescriber(modifier);
+    }
+}
+
+/**
+ * Inserts a prescriber Prescription to the db. If modifier is empty then inserts genericPrescriberPrescription.
+ * Else, overwrites the specified fields in genericPrescriberPrescription with the values in 
+ * modifier then inserts the modified prescriberPrescription.
+ * 
+ * Returns the prescriberPrescription that was inserted.
+ * @param {Object} modifier optional object to overwrite values in genericPrescriberPrescription .
+ * @returns {Object} the prescriberPrescription that was inserted.
+ */
+export const insertPrescriberFrom = async (modifier = {}) => {
+    let prescriberPrescription = await objWithModifier(genericPrescriberPrescription, modifier);
+    await db.collection(COLLECTIONS.PRESCRIBER_PRESCRIPTIONS).insertOne(prescriberPrescription);
+    return prescriberPrescription;
+}
+
+/**
+ * Insert numPrescribersPrescription prescribers into the db. 
+ * Each prescriber Prescription is generated from the generic prescriber Prescription Prescriptionat
+ * with incrementing date "2024-12-{i}" and 
+ * padded providerCode "ON-JC001"
+ * @param {number} numPrescriberPrescriptions 
+ */
+export const insertPrescriberPrescriptions = async (numPrescriberPrescriptions = 20) => {
+    for (let i = 1; i <= numPrescriberPrescriptions; i++) {
+        const modifier = {
+            date: `2024-12-${i}`
+        }
+        await insertPrescriberFrom(modifier);
+    }
+}
+
 
 /**
  * Returns a clone of obj with the KVP's specified in opts
